Alterações em UsuarioService, organizações e alteração no PessoaDAO

+++ app/src/main/java/com/ufrpe/feelingsbox/usuario/persistencia/UsuarioDAO.java
@@ -88,7 +88,7 @@ public class UsuarioDAO {
     }
 
 
-    public Usuario getUsuarioId(int id){
+    public Usuario getUsuarioId(long id){
         feelingsDb = dbHelper.getReadableDatabase();
 
         String query = "SELECT * FROM " + DataBase.TABELA_USUARIO +
